requirements.txt
# Auto generated by pip-freeze. DO NOT EDIT THIS FILE.

# These are the requirements for the main code bundle.
# It excludes things needed to build lava but not needed to run it.

# ex: set ft=sh:

boto3==1.39.16
botocore==1.39.16
cachetools==6.1.0
colorama==0.4.6
cookiecutter==2.6.0
cx_Oracle==8.3.0
docker==7.1.0
jinja2==3.1.6
jsonschema==4.25.0
pg8000==1.31.4
psutil==7.0.0
pygments==2.19.2
pymysql==1.1.1
pysmb==1.2.11
python-dateutil==2.9.0.post0
pyyaml==6.0.2
requests==2.32.4
smbprotocol==1.15.0
sqlalchemy==2.0.42
sqlparse==0.5.3
graphlib-backport>=1.0.3; python_version < '3.9'
tabulate==0.9.0
smart_open[s3]==7.3.0.post1
packaging==25.0
